Handle html returned as ajax response
The responses to ajax requests in qr.js and citation.js are not expected
to be html. It seems there are cases where titles that do not exist in
the crossref & datacite databases resolve to the DOI destination instead
of returning a failed metadata notice. Treat these cases as failed
fetches.