Fix DOI title features
A recent Chrome release changed how/where permissions may be requested.
It is no longer possible to maintain a single method in the background
that can be called from other scripts. Move setDoiMetaPermissions to
each script where permissions are requested and call it locally.

When recording a DOI from the bubble input, the window closes before
processing has finished, often resulting in failed DOI recording.
Implement a queue that can be written to quickly (and awaited) before
closing the bubble and process the queue separately in the background.