Update storage change listener for cross browser compatibility
Firefox does not filter change reports to only those changes whose value
actually updated. When options are set in a storage area, all of those
options are reported as changed. Manually filter changes and cross-area
updates to only include values that actually changed.

Improve the capabilities of storageListenerStatus so that it is possible
to skip the next update, whenever it occurs over the next 100ms, rather
than trusting that the change listener will fire before the promise from
setting options in a storage area resolves.